Src folder created and tictactoe files moved there
