+++ app.js
@@ -1,2 +1,2 @@
-const tictactoe = require("./tictactoe");
+const tictactoe = require("./src/tictactoe");
 console.log(tictactoe("Tic Tac Toe"));
